ENG-119 Move Search criteria to ArticleSearchQuery
