+++ src/main/java/org/ambraproject/wombat/controller/HomeController.java
@@ -90,13 +90,13 @@ public class HomeController extends WombatController {
     RECENT {
       @Override
       public List<SolrArticleAdapter> getArticles(HomeController context, SectionSpec section, Site site, int start) throws IOException {
-        return getArticlesFromSolr(context, section, site, start, SolrSearchApi.SolrSortOrder.DATE_NEWEST_FIRST);
+        return getArticlesFromSolr(context, section, site, start, ArticleSearchQuery.SolrSortOrder.DATE_NEWEST_FIRST);
       }
     },
     POPULAR {
       @Override
       public List<SolrArticleAdapter> getArticles(HomeController context, SectionSpec section, Site site, int start) throws IOException {
-        return getArticlesFromSolr(context, section, site, start, SolrSearchApi.SolrSortOrder.MOST_VIEWS_30_DAYS);
+        return getArticlesFromSolr(context, section, site, start, ArticleSearchQuery.SolrSortOrder.MOST_VIEWS_30_DAYS);
       }
     },
     CURATED {
@@ -134,14 +134,14 @@ public class HomeController extends WombatController {
     };
 
     private static List<SolrArticleAdapter> getArticlesFromSolr(HomeController context, SectionSpec section, Site site, int start,
-                                                                SolrSearchApi.SolrSortOrder order)
+                                                                ArticleSearchQuery.SolrSortOrder order)
         throws IOException {
       ArticleSearchQuery query = ArticleSearchQuery.builder()
         .setStart(start)
         .setRows(section.resultCount)
         .setSortOrder(order)
         .setJournalKeys(ImmutableList.of(site.getJournalKey()))
-        .setDateRange(SolrSearchApi.SolrEnumeratedDateRange.ALL_TIME)
+        .setDateRange(ArticleSearchQuery.SolrEnumeratedDateRange.ALL_TIME)
         .build();
       return SolrArticleAdapter.unpackSolrQuery(context.solrSearchApi.search(query));
     }
@@ -329,9 +329,9 @@ public class HomeController extends WombatController {
     ArticleSearchQuery query = ArticleSearchQuery.builder()
         .setStart(0)
         .setRows(getFeedLength(site))
-        .setSortOrder(SolrSearchApi.SolrSortOrder.DATE_NEWEST_FIRST)
+        .setSortOrder(ArticleSearchQuery.SolrSortOrder.DATE_NEWEST_FIRST)
         .setJournalKeys(ImmutableList.of(site.getJournalKey()))
-        .setDateRange(SolrSearchApi.SolrEnumeratedDateRange.ALL_TIME)
+        .setDateRange(ArticleSearchQuery.SolrEnumeratedDateRange.ALL_TIME)
       .setRssSearch(true).build();
     SolrSearchApi.Result recentArticles = solrSearchApi.search(query);
