+++ src/main/java/org/ambraproject/wombat/service/remote/SolrSearchApi.java
@@ -29,16 +29,10 @@ import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
-import java.util.TimeZone;
-import javax.xml.bind.DatatypeConverter;
 import com.google.auto.value.AutoValue;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
@@ -158,139 +152,6 @@ public class SolrSearchApi {
           .build();
       }
     }
-  }
-
-  /**
-   * Enumerates sort orders that we want to expose in the UI.
-   */
-  public static enum SolrSortOrder implements ArticleSearchQuery.SearchCriterion {
-
-    // The order here determines the order in the UI.
-    RELEVANCE("Relevance", "score desc,publication_date desc"),
-    DATE_NEWEST_FIRST("Date, newest first", "publication_date desc"),
-    DATE_OLDEST_FIRST("Date, oldest first", "publication_date asc"),
-    MOST_VIEWS_30_DAYS("Most views, last 30 days", "counter_total_month desc"),
-    MOST_VIEWS_ALL_TIME("Most views, all time", "counter_total_all desc"),
-    MOST_CITED("Most cited, all time", "alm_scopusCiteCount desc"),
-    MOST_BOOKMARKED("Most bookmarked", "alm_mendeleyCount desc"),
-    MOST_SHARED("Most shared in social media", "sum(alm_twitterCount, alm_facebookCount) desc");
-
-    private String description;
-
-    private String value;
-
-    SolrSortOrder(String description, String value) {
-      this.description = description;
-      this.value = value;
-    }
-
-    @Override
-    public String getDescription() {
-      return description;
-    }
-
-    @Override
-    public String getValue() {
-      return value;
-    }
-  }
-
-  /**
-   * Enumerates date ranges to expose in the UI.  Currently, these all start at some prior date and extend to today.
-   */
-  public static enum SolrEnumeratedDateRange implements ArticleSearchQuery.SearchCriterion {
-
-    ALL_TIME("All time", -1),
-    LAST_YEAR("Last year", 365),
-
-    // Clearly these are approximations given the different lengths of months.
-    LAST_6_MONTHS("Last 6 months", 182),
-    LAST_3_MONTHS("Last 3 months", 91);
-
-    private String description;
-
-    private int daysAgo;
-
-    SolrEnumeratedDateRange(String description, int daysAgo) {
-      this.description = description;
-      this.daysAgo = daysAgo;
-    }
-
-    @Override
-    public String getDescription() {
-      return description;
-    }
-
-    /**
-     * @return a String representing part of the "fq" param to pass to solr that will restrict the date range
-     * appropriately.  For example, "[2013-02-14T21:00:29.942Z TO 2013-08-15T21:00:29.942Z]". The String must be escaped
-     * appropriately before being included in the URL.  The final http param passed to solr should look like
-     * "fq=publication_date:[2013-02-14T21:00:29.942Z+TO+2013-08-15T21:00:29.942Z]". If this date range is ALL_TIME,
-     * this method returns null.
-     */
-    @Override
-    public String getValue() {
-      if (daysAgo > 0) {
-        Calendar today = Calendar.getInstance();
-        today.setTimeZone(TimeZone.getTimeZone("UTC"));
-        Calendar then = Calendar.getInstance();
-        then.setTimeZone(TimeZone.getTimeZone("UTC"));
-        then.add(Calendar.DAY_OF_YEAR, -daysAgo);
-        return String.format("[%s TO %s]", DatatypeConverter.printDateTime(then),
-            DatatypeConverter.printDateTime(today));
-      } else {
-        return null;
-      }
-    }
-  }
-
-  public static class SolrExplicitDateRange implements ArticleSearchQuery.SearchCriterion {
-
-    private String description;
-    private Calendar startDate;
-    private Calendar endDate;
-
-    public SolrExplicitDateRange(String description, String startDate, String endDate) {
-      this.description = description;
-
-      Calendar startCal = Calendar.getInstance();
-      Calendar endCal = Calendar.getInstance();
-      startCal.setTimeZone(TimeZone.getTimeZone("UTC"));
-      endCal.setTimeZone(TimeZone.getTimeZone("UTC"));
-
-      SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd", Locale.US);
-      // getValue() method uses DatatypeConverter.printDateTime to convert the calendar object to a string.
-      // However, this method uses the local time zone. Setting the time zone for the Calendar object doesn't
-      // enforce UTC in the result of the printDateTime method but setting it in the simpleDateFormat does.
-      simpleDateFormat.setTimeZone(TimeZone.getTimeZone("UTC"));
-      try {
-        startCal.setTime(simpleDateFormat.parse(startDate));
-        endCal.setTime(simpleDateFormat.parse(endDate));
-      } catch (ParseException e) {
-        throw new RuntimeException(e);
-      }
-
-      this.startDate = startCal;
-      this.endDate = endCal;
-    }
-
-    @Override
-    public String getDescription() {
-      return description;
-    }
-
-    /**
-     * @return a String representing part of the "fq" param to pass to solr that will restrict the date range
-     * appropriately.  For example, "[2013-02-14T21:00:29.942Z TO 2013-08-15T21:00:29.942Z]". The String must be escaped
-     * appropriately before being included in the URL.  The final http param passed to solr should look like
-     * "fq=publication_date:[2013-02-14T21:00:29.942Z+TO+2013-08-15T21:00:29.942Z]".
-     */
-    @Override
-    public String getValue() {
-      return String.format("[%s TO %s]", DatatypeConverter.printDateTime(startDate),
-          DatatypeConverter.printDateTime(endDate));
-    }
-
   }
 
   /**
